+++ product-cateogory-crud/src/main/java/com/nimapinfotech/product_cateogory_crud/controller/ProductController.java
@@ -5,6 +5,7 @@ import com.nimapinfotech.product_cateogory_crud.service.ProductService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.Optional;
@@ -21,24 +22,49 @@ public class ProductController {
         return productService.getAllProducts(pageable);
     }
 
+    //get all prodcuts
     @PostMapping("/products")
     public Product createProduct(@RequestBody Product product) {
         return productService.createProduct(product);
     }
 
+
     @GetMapping("products/{id}")
-    public Optional<Product> getProductById(@PathVariable Long id) {
-        return productService.getProductById(id);
+    public ResponseEntity<Product> getProductById(@PathVariable Long id) {
+         Product product = productService.getProductById(id).orElse(null);
+         if (product!=null){
+             return ResponseEntity.ok().body(product);
+         }
+         else {
+             return ResponseEntity.notFound().build();
+         }
     }
 
     @PutMapping("products/{id}")
-    public Product updateProductById(@PathVariable Long id,@RequestBody Product updatedProduct){
-        return productService.updateProductById(id,updatedProduct);
+    public ResponseEntity<Product> updateProductById(@PathVariable Long id,@RequestBody Product updatedProduct){
+        Product product = productService.getProductById(id).orElse(null);
+        if (product!=null){
+            Product newproduct= productService.updateProductById(id,updatedProduct);
+            return ResponseEntity.ok().body(newproduct);
+        }
+        else {
+            return ResponseEntity.notFound().build();
+        }
+
     }
 
     @DeleteMapping("/products/{id}")
-    public void deleteProductById(@PathVariable Long id){
-        productService.deleteProductById(id);
+    public ResponseEntity<Product> deleteProductById(@PathVariable Long id){
+        Product product = productService.getProductById(id).orElse(null);
+        if (product!=null){
+            productService.deleteProductById(id);
+
+            return ResponseEntity.ok().body(product);
+        }
+        else {
+            return ResponseEntity.notFound().build();
+        }
+
     }
 
 }
